requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
Werkzeug==2.3.4

pip~=23.2.1
typing_extensions~=4.12.2
MarkupSafe~=3.0.1
click~=8.1.7
blinker~=1.8.2
Jinja2~=3.1.4
itsdangerous~=2.2.0
visitor~=0.1.3
SQLAlchemy~=2.0.35
python-dotenv~=1.0.1
Mako~=1.3.5
alembic~=1.13.3
Flask-Migrate~=4.0.7
gunicorn